🐛 Change setLogin state from page template -> App
- page template을 쓰지 않는 페이지도 있기 때문
