+++ src/Template/PageTemplate/index.tsx
@@ -1,7 +1,4 @@
-import { useEffect } from 'react';
 import { Header, Footer } from '../';
-import { useAuthStore } from '../../hooks/useStore';
-import { getUserInfo } from '../../utils/userInfo';
 import { mainStyle } from './style.css';
 
 interface IPageTemplate {
@@ -9,15 +6,6 @@ interface IPageTemplate {
 }
 
 function PageTemplate({ children }: IPageTemplate) {
-  const { setIsLogin } = useAuthStore();
-
-  useEffect(() => {
-    const userInfo = getUserInfo();
-    if (userInfo) {
-      setIsLogin(true);
-    }
-  }, []);
-
   return (
     <>
       <Header />
